Big commit -Refactored structure of app to get babel working -Found blend of plugins and  configs to get babel production build working
