+++ .babelrc
@@ -1,5 +1,5 @@
 {
 	"presets": [
 		"@babel/preset-env"
-	]
+  ]
 }
