+++ package.json
@@ -7,20 +7,25 @@
     "@babel/cli": "^7.6.0",
     "@babel/core": "^7.6.0",
     "@babel/node": "^7.6.1",
+    "@babel/polyfill": "^7.6.0",
     "@babel/preset-env": "^7.6.0",
     "@types/dotenv": "^6.1.1",
     "cloudinary": "^1.15.0",
+    "core-js": "^3.2.1",
     "cors": "^2.8.5",
     "dotenv": "^8.1.0",
     "express": "^4.17.1",
+    "express-async-errors": "^3.1.1",
     "express-form-data": "^2.0.10",
     "knex": "^0.19.4",
     "knex-cleaner": "^1.3.0",
     "pg": "^7.12.1",
+    "regenerator-runtime": "^0.13.3",
     "uuid": "^3.3.3"
   },
   "description": "",
   "devDependencies": {
+    "@babel/plugin-transform-async-to-generator": "^7.5.0",
     "nodemon": "^1.19.2"
   },
   "engines": {
@@ -37,8 +42,8 @@
   },
   "scripts": {
     "server": "nodemon --exec babel-node src/index.js",
-    "build": "npm run clean && npm run build-babel",
-    "start": "npm run build && node ./build/index.js",
+    "build": "yarn clean && yarn build-babel",
+    "start": "yarn build && node ./build/index.js",
     "clean": "rm -rf build && mkdir build",
     "build-babel": "babel -d ./build ./src -s",
     "test": "echo \"Error: no test specified\" && exit 1"
