+++ src/data/dbConfig.js
@@ -1,7 +1,7 @@
-require('dotenv').config();
+//require('dotenv').config();
 const dbEngine = process.env.DB_ENV || "development";
 import knex from 'knex';
-import knexConfigs from '../knexfile';
+import knexConfigs from '../../knexfile';
 const config = knexConfigs[dbEngine];
 
 console.log(config);
